% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_matrix_normal.R
\name{fast_dmatnorm}
\alias{fast_dmatnorm}
\title{Calculate PDF for the matrix normal distribution accepting both covariance and precision}
\usage{
fast_dmatnorm(Z, M, U, V, log = TRUE, Precision = FALSE, tol = 1e-08)
}
\arguments{
\item{Z}{The number of observations to be sampled}

\item{M}{The Mean matrix}

\item{U}{The row covariance/precision matrix with dimension n * n}

\item{V}{The column covariance/precision matrix with dimension p * p}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}

\item{Precision}{logical; if TRUE U and V will be considered as precision}

\item{tol}{a small number with default number 1e-6 which is used if the the full rank precision issue encountered}
}
\value{
Return the pdf of Z
}
\description{
Calculate PDF for the matrix normal distribution accepting both covariance and precision
}
\examples{
n = 100
p = 100
U <- matrix(0.5,nrow=n,ncol=n) + 0.5*diag(n)
V <- matrix(0.8,nrow=p,ncol=p) + 0.2*diag(p)
Y<- fast_rmatnorm(n = n, p = p, U_cov = U, V_cov=V)
M = matrix(0, nrow = n, ncol = p)
p_val = fast_dmatnorm(Y, M, U, V)
}
