% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_matrix_normal.R
\name{fast_rmatnorm}
\alias{fast_rmatnorm}
\title{Random generation for the matrix normal distribution accepting both covariance and precision}
\usage{
fast_rmatnorm(
  num_samp = 1,
  n = 10,
  p = 5,
  M = NULL,
  U_cov = NULL,
  V_cov = NULL,
  U_prec = NULL,
  V_prec = NULL,
  useCov = TRUE,
  small_tol = 1e-06
)
}
\arguments{
\item{num_samp}{The number of observations to be sampled}

\item{n}{The row dimension of the sampled matrix normal variable}

\item{p}{The column dimension of the sampled matrix normal variable}

\item{M}{The Mean matrix}

\item{U_cov}{The row covariance matrix with dimension n * n}

\item{V_cov}{The column covariance matrix with dimension p * p}

\item{U_prec}{The row precision matrix with dimension n * n}

\item{V_prec}{The column precision matrix with dimension p * p}

\item{useCov}{logical; if TRUE the U_cov and V_cov will be used otherwise U_prec and V_prec will be used}

\item{small_tol}{a small number with default number 1e-6 which is used if the the full rank precision issue encountered}
}
\value{
If num_samp = 1, it will return a matrix. Otherwise a 3-D array will be returned
}
\description{
Random generation for the matrix normal distribution accepting both covariance and precision
}
\examples{
n = 100
p = 100
U <- matrix(0.5,nrow=n,ncol=n) + 0.5*diag(n)
V <- matrix(0.8,nrow=p,ncol=p) + 0.2*diag(p)
Y<- fast_rmatnorm(U_cov = U, V_cov=V)
}
